requirements.txt
google-generativeai>=0.8.0
langgraph>=0.2.0
neo4j>=5.0.0
streamlit>=1.40.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
plotly>=5.17.0
